requirements.txt
amqp==5.1.1
ansicon==1.89.0
APScheduler==3.9.1.post1
arrow==1.2.3
asgiref==3.6.0
billiard==3.6.4.0
blessed==1.19.1
bokeh==3.0.3
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.5
django-apscheduler==0.3.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cronjobs==0.2.3
django-crontab==0.7.1
django-picklefield==3.1
django-q==1.3.9
django-q2==1.4.9
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
postmarker==1.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
xyzservices==2022.9.0
